+++ visitz/Visitz/Pages/AppLockPage.xaml
@@ -1,10 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
-<views:VisitzPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+<pages:VisitzPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:views="clr-namespace:Visitz.Views"
+             xmlns:pages="clr-namespace:Visitz.Pages"
              xmlns:viewmodels="clr-namespace:Visitz.ViewModels"
              xmlns:local="clr-namespace:Visitz.Localization"
-             x:Class="Visitz.Views.AppLockPage"
+             x:Class="Visitz.Pages.AppLockPage"
              x:DataType="viewmodels:AppLockViewModel"
              Title=""
              BackgroundColor="{StaticResource Primary}">
@@ -47,4 +47,4 @@
 
     </VerticalStackLayout>
 
-</views:VisitzPage>
+</pages:VisitzPage>
