+++ visitz/Visitz/Pages/AppLockPage.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using Visitz.ViewModels;
 
-namespace Visitz.Views;
+namespace Visitz.Pages;
 
 public partial class AppLockPage : VisitzPage
 {
