+++ visitz/Visitz/Pages/CaseloadItemDetailsPage.xaml
@@ -1,11 +1,11 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
-<views:VisitzPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+<pages:VisitzPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
              xmlns:controls="clr-namespace:Visitz.Controls"
-             xmlns:views="clr-namespace:Visitz.Views"
+             xmlns:pages="clr-namespace:Visitz.Pages"
              xmlns:models="clr-namespace:Visitz.Models"
              xmlns:viewModels="clr-namespace:Visitz.ViewModels"
-             x:Class="Visitz.Views.CaseloadItemDetailsPage"
+             x:Class="Visitz.Pages.CaseloadItemDetailsPage"
              x:DataType="viewModels:CaseloadItemDetailsViewModel"
              Title="Details">
     
@@ -138,4 +138,4 @@
         </VerticalStackLayout>
 
     </ScrollView>
-</views:VisitzPage>
+</pages:VisitzPage>
