+++ visitz/Visitz/Pages/CaseloadPage.xaml
@@ -1,11 +1,11 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
-<views:VisitzPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+<pages:VisitzPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:views="clr-namespace:Visitz.Views"
+             xmlns:pages="clr-namespace:Visitz.Pages"
              xmlns:viewmodels="clr-namespace:Visitz.ViewModels"
              xmlns:models="clr-namespace:Visitz.Models"
              xmlns:controls="clr-namespace:Visitz.Controls"
-             x:Class="Visitz.Views.CaseloadPage"
+             x:Class="Visitz.Pages.CaseloadPage"
              x:DataType="viewmodels:CaseloadViewModel"
              Title="Caseload">
 
@@ -62,4 +62,4 @@
 
     </RefreshView>
 
-</views:VisitzPage>
+</pages:VisitzPage>
