+++ visitz/Visitz/Pages/DebugOptionsPage.xaml
@@ -1,9 +1,9 @@
 <?xml version="1.0" encoding="utf-8" ?>
-<views:VisitzPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+<pages:VisitzPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
                   xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-                  xmlns:views="clr-namespace:Visitz.Views"
+                  xmlns:pages="clr-namespace:Visitz.Pages"
                   xmlns:viewmodels="clr-namespace:Visitz.ViewModels"
-                  x:Class="Visitz.Views.DebugOptionsPage"
+                  x:Class="Visitz.Pages.DebugOptionsPage"
                   x:DataType="viewmodels:DebugOptionsViewModel"
                   Title="Debug Options">
 
@@ -82,4 +82,4 @@
 
     </Grid>
 
-</views:VisitzPage>
+</pages:VisitzPage>
