+++ visitz/Visitz/Pages/NotesPage.xaml
@@ -1,10 +1,10 @@
 ﻿<?xml version="1.0" encoding="utf-8" ?>
-<views:VisitzPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
+<pages:VisitzPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             xmlns:views="clr-namespace:Visitz.Views"
+             xmlns:pages="clr-namespace:Visitz.Pages"
              xmlns:viewModels="clr-namespace:Visitz.ViewModels"
              xmlns:models="clr-namespace:Visitz.Models"
-             x:Class="Visitz.Views.NotesPage"
+             x:Class="Visitz.Pages.NotesPage"
              x:DataType="viewModels:NotesViewModel"
              Title="Notes">
 
@@ -143,5 +143,5 @@
 
         </Button>
     </Grid>
-</views:VisitzPage>
+</pages:VisitzPage>
