+++ visitz/Visitz/Pages/NotesPage.xaml.cs
@@ -1,6 +1,6 @@
 ﻿using Visitz.ViewModels;
 
-namespace Visitz.Views;
+namespace Visitz.Pages;
 
 public partial class NotesPage : VisitzPage
 {
