+++ visitz/Visitz/Pages/VisitzPage.xaml
@@ -2,7 +2,7 @@
 <!--The base page xaml for all the other xamls. Common functionality can be defined here.-->
 <ContentPage xmlns="http://schemas.microsoft.com/dotnet/2021/maui"
              xmlns:x="http://schemas.microsoft.com/winfx/2009/xaml"
-             x:Class="Visitz.Views.VisitzPage"
+             x:Class="Visitz.Pages.VisitzPage"
              Title="VisitzPage">
 </ContentPage>
