+++ visitz/Visitz/VisitzApp.xaml.cs
@@ -1,4 +1,4 @@
-﻿using Visitz.Views;
+﻿using Visitz.Pages;
 
 namespace Visitz;
